Move main() to a separated file, version definition to version.h
Moved the main function to a separated file so we can in the future use the same
main() for different views/types of window.
The ICEBERG_VERSION definition was moved from mainwindow.cpp to version.h.
