+++ src/version.h
@@ -1,0 +1,6 @@
+#ifndef VERSION_H
+#define VERSION_H
+
+#define ICEBERG_VERSION "1.0"
+
+#endif // VERSION_H
